refactor: fix imports after dataset rename, clarify config: prefix in CLI help, add config-diagram pre-commit hook
