+++ src/mlcast/__main__.py
@@ -131,15 +131,19 @@ def _build_help_text(cfg: fdl.Buildable) -> Text:
         t.append("\n")
 
     t.append("\nSwitching experiments:\n", style="bold yellow")
-    t.append("  # Switch to a completely different base config function (if defined):\n", style="dim")
-    t.append("  mlcast train ", style="bold")
-    t.append("--config=config:another_experiment_function\n", style="cyan")
     t.append("\n  # Resume from or reproduce a previously saved YAML config:\n", style="dim")
     t.append("  mlcast train ", style="bold")
     t.append("--config /path/to/config.yaml\n", style="cyan")
     t.append("\n  # Load a YAML config and apply additional overrides on top:\n", style="dim")
     t.append("  mlcast train ", style="bold")
     t.append("--config /path/to/config.yaml --config set:trainer.max_epochs=50\n", style="cyan")
+    t.append("\n  # Use a different base config function defined in src/mlcast/config/.\n", style="dim")
+    t.append("  # Syntax: --config=config:<function_name>  where ", style="dim")
+    t.append("config:", style="dim bold")
+    t.append(" is a Fiddle prefix that resolves\n", style="dim")
+    t.append("  # the function name against the config module (not a Python module path).\n", style="dim")
+    t.append("  mlcast train ", style="bold")
+    t.append("--config=config:another_experiment_function\n", style="cyan")
 
     t.append("\nInspecting the resolved config:\n", style="bold yellow")
     t.append("  # Print the fully resolved config as YAML without starting training:\n", style="dim")
