+++ src/mlcast/data/__init__.py
@@ -1,4 +1,4 @@
 from .source_data_datamodule import SourceDataDataModule
-from .source_datasets import SourceDataPrecomputedSamplingDataset
+from .source_data_datasets import SourceDataPrecomputedSamplingDataset
 
 __all__ = ["SourceDataDataModule", "SourceDataPrecomputedSamplingDataset"]
